Refactored routes folder folder
