+++ src/components/app/app.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {appPropertiesRequest} from "../../redux/appProperties/actions";
 import Header from '../header';
 import NavigationBar from "../navigation_bar";
-import Routes from "../routes/routes";
+import Routes from "../../routes/routes";
 
 const App = ({store}) => {
     store.dispatch(appPropertiesRequest());
